biokeeper-auth/requirements_test.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements_test.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.1.3
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via
    #   typer
    #   uvicorn
cryptography==42.0.7
databases==0.9.0
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via fastapi
exceptiongroup==1.2.1
    # via
    #   anyio
    #   pytest
fastapi==0.111.0
fastapi-cli==0.0.4
    # via fastapi
greenlet==3.0.3
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
jinja2==3.1.4
    # via fastapi
load-dotenv==0.1.0
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
orjson==3.10.3
    # via fastapi
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.9
pycparser==2.22
    # via cffi
pydantic==2.7.2
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.18.3
    # via pydantic
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pygments==2.18.0
    # via rich
pyjwt==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
    # via
    #   load-dotenv
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via uvicorn
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.30
    # via databases
starlette==0.37.2
    # via fastapi
tomli==2.0.1
    # via pytest
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   uvicorn
ua-parser==0.18.0
    # via user-agents
ujson==5.10.0
    # via fastapi
user-agents==2.2.0
uvicorn==0.30.0
    # via fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
